+++ src/components/index.ts
@@ -8,6 +8,7 @@ export { MacroStrategyCard } from './cards/MacroStrategyCard';
 export { DecisionLog } from './cards/DecisionLog';
 export { MarketOverview } from './cards/MarketOverview';
 export { SystemStatus } from './cards/SystemStatus';
+export { CostSummaryCard } from './cards/CostSummaryCard';
 
 // Charts
 export { EquityCurve } from './charts/EquityCurve';